fix(sentry): recognise the contract's own security reason; extract the lookups
